fix(verification): break circular import via dedicated _types module
Extract CheckResult into github_operations/_types.py so both
verification.py and _permission_probes.py can import it without
re-entering verification during its own load phase.

The previous fix (moving `logger = ...` after CheckResult) did not
actually break the cycle — the `from ._permission_probes import`
statement remained above the CheckResult definition, so loading
_permission_probes still re-entered verification before CheckResult
was bound. After the rebase onto current main, the cycle resurfaced
and broke every test that touched the github_operations package.

verification.py re-exports CheckResult via __all__ so existing
callers (github_operations/__init__.py, etc.) keep working
unchanged.
