+++ src/mcp_workspace/github_operations/verification.py
@@ -6,7 +6,7 @@ and returns per-check results. All checks use PyGithub — no gh CLI dependency.
 
 import logging
 from pathlib import Path
-from typing import Any, Literal, NotRequired, TypedDict
+from typing import Any
 
 from github import Auth, Github
 from github.GithubException import GithubException
@@ -15,27 +15,12 @@ from mcp_coder_utils.user_app_data import get_user_app_data_dir
 from mcp_workspace.config import get_github_token_with_source
 from mcp_workspace.git_operations.remotes import get_repository_identifier
 from mcp_workspace.github_operations._diagnostics import extract_diagnostic_headers
-
-# Imported after CheckResult is defined to break the circular import:
-# `_permission_probes` imports CheckResult from this module.
-from mcp_workspace.github_operations._permission_probes import (  # noqa: E402  # pylint: disable=wrong-import-position
-    run_permission_probes,
-)
+from mcp_workspace.github_operations._permission_probes import run_permission_probes
+from mcp_workspace.github_operations._types import CheckResult
 from mcp_workspace.github_operations.base_manager import BaseGitHubManager
 from mcp_workspace.utils.token_fingerprint import format_token_fingerprint
 
-
-class CheckResult(TypedDict):
-    """Result of a single verification check."""
-
-    ok: bool
-    value: str
-    severity: Literal["error", "warning"]
-    error: NotRequired[str]
-    install_hint: NotRequired[str]
-    token_source: NotRequired[Literal["env", "config"]]
-    token_fingerprint: NotRequired[str]
-
+__all__ = ["CheckResult", "verify_github"]
 
 logger = logging.getLogger(__name__)
